Baranches merged

+++ src/SalesManager.java
@@ -14,10 +14,10 @@ public class SalesManager {
         }
         return max;
     }
-    public int trimAverage() {
-        int max = 0;
-        int min = sales[0];
-        int sum = 0;
+    public long trimAverage() {
+        long max = 0;
+        long min = sales[0];
+        long sum = 0;
         for (int i = 1; i < sales.length; i++) {
             if (sales[i] <= min) {
                 sum += min;
